@@ -1,35 +1,34 @@
 {
   "name": "yawm-frontend",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.8.2",
+    "@mui/material": "^5.8.1",
     "@sentry/react": "^6.19.7",
     "@sentry/tracing": "^6.19.7",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@webscopeio/react-health-check": "^3.0.0",
     "axios": "^0.24.0",
-    "bootstrap": "^5.1.3",
+    "date-fns": "^2.28.0",
     "js-cookie": "^3.0.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.6.4",
-    "react-bootstrap-icons": "^1.5.0",
-    "react-dom": "^17.0.2",
-    "react-google-login": "^5.2.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^4.0.3",
-    "web-vitals": "^1.1.2"
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "coverage": "react-scripts test --coverage"
   },
   "eslintConfig": {
     "extends": [
-      "react-app"
+      "react-app",
+      "react-app/jest"
     ]
   },
   "browserslist": {
@@ -43,5 +42,23 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.4",
+    "eslint": "^8.15.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^26.5.3",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.29.4",
+    "husky": "^8.0.1",
+    "lint-staged": "^12.4.1",
+    "prettier": "^2.6.2",
+    "pretty-quick": "^3.1.3",
+    "react-test-renderer": "^18.2.0",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^13.5.0"
   }
 }
