@@ -1,24 +1,26 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./components/App";
-import "bootstrap/dist/css/bootstrap.min.css";
-import * as Sentry from "@sentry/react";
-import { BrowserTracing } from "@sentry/tracing";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import * as Sentry from '@sentry/react';
+import { BrowserTracing } from '@sentry/tracing';
+import App from './components/App';
 
-const { REACT_APP_SENTRY_DSN } = process.env;
+const DEBUG = process.env.REACT_APP_DEBUG;
 
 Sentry.init({
-    debug: true,
-    dsn: REACT_APP_SENTRY_DSN,
-    integrations: [new BrowserTracing()],
-    // We recommend adjusting this value in production, or using tracesSampler
-    // for finer control
-    tracesSampleRate: 1.0,
+  debug: DEBUG || false,
+  release: 'yawm-frontend@0.3.0',
+  environment: DEBUG ? 'dev' : 'production',
+  // JS can`t securely store DSN, so, there is no need to hide it
+  dsn: 'https://c7c50c25cc3448c3b7f9da3915c05110@o1201644.ingest.sentry.io/6373679',
+  integrations: [new BrowserTracing()],
+  // We recommend adjusting this value in production, or using tracesSampler
+  // for finer control
+  tracesSampleRate: 1.0,
 });
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
